requirements.txt
transformers
accelerate
python-dotenv
datasets
torch
sympy
google-generativeai